requirements.txt
Django==5.2.9
djangorestframework==3.16.1
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.5
pandas==2.2.3
langchain-openai==0.2.14
crewai==1.7.1
crewai-tools==1.7.1
openlit==1.36.3
pydantic==2.11.10
pydantic-settings==2.10.1
